Change move-to feature - files path, rename variables and files
Feature #14103

Arvados-DCO-1.1-Signed-off-by: Janicki Artur <artur.janicki@contractors.roche.com>
